project: rename sources/headers with too long name
- tag_duplicates is now `dedupe`
- user_input is now `uinput`
